We got rid of an extra static import that wasnt needed and fixed methods

+++ src/main/java/frc/robot/Constants.java
@@ -23,7 +23,7 @@ public class Constants {
         public static double kI;
         public static double kD;
 
-        public static double currentLimit;
+        public static int currentLimit;
         public static int motor1;
     }
 }
